Factor sphere drawing into a separate "renderer" class
