+++ src/Application/UI/Simple.cpp
@@ -4,6 +4,7 @@ UI::Simple::Simple(NBody::Simulation &simulation) :
         Gtk::ApplicationWindow(),
         _simulation{simulation},
         _camera{
+                simulation,
                 Vector3::zAxis(-20.0f),
                 {}, Vector3::yAxis(),
                 45.0_degf,
