+++ src/NBody/View/CMakeLists.txt
@@ -30,7 +30,9 @@ add_library(View
         GtkmmArcBallCamera.cpp GtkmmArcBallCamera.h
         InteractiveView.cpp InteractiveView.h
         View.cpp View.h
-        )
+        Renderer.h
+        Renderers/InstancedPhongRenderer.cpp Renderers/InstancedPhongRenderer.h
+        Renderers/PhongRenderer.cpp Renderers/PhongRenderer.h)
 
 target_include_directories(View PUBLIC
         ${CMAKE_SOURCE_DIR}/src)
